Times pending for changes

+++ src/components/Products/index.jsx
@@ -1,22 +1,34 @@
 import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import ProductCard from '../ProductCard';
-import getAllProducts from '../../store/actions';
+import { getAllProducts, setInitialTimes } from '../../store/actions';
 import './styles.scss';
 import Spinner from '../../imgs/Spin.svg';
 
 const Products = () => {
   const products = useSelector((state) => state.products);
   const loading = useSelector((state) => state.loading);
+
   const dispatch = useDispatch();
 
+  if (!localStorage.length && products) {
+    products.forEach((product) => {
+      localStorage.setItem(product.id, Math.floor(Math.random() * 100) + 60);
+    });
+  }
   useEffect(() => {
     const fetchProducts = async () => {
-      localStorage.clear();
-      getAllProducts(dispatch);
+      if (products.length < 1) {
+        localStorage.clear();
+        getAllProducts(dispatch);
+      }
     };
     fetchProducts();
   }, []);
+
+  if (products.length > 1) {
+    setInitialTimes(dispatch, products);
+  }
 
   return (
     <div className="Products">
